requirements.txt
dbt-postgres==1.10.0
dbt-core==1.11.6
psycopg[binary]>=3,<4
python-dotenv>=1,<2
azure-servicebus>=7,<8
